chore: 파일경로수정 및 리뷰반영
