+++ src/lib/utils/percentageCalculator.ts
